38.3....Object Rest Spread{Installing & Configuring babel-plugin-transform-object-rest-spread: yarn add babel-plugin-transform-object-rest-spread --dev}
Or does the "spread" syntax work for objects? Google for "babel plugin transform object rest spread" and find the Babel documentation page. The feature we're "dreaming" about is called "object rest spread". It is not an official ECMAScript feature. But, it's currently a proposed, "draft" feature that's in a late stage. There's no promises, but that means it will likely become a real feature in a future ECMAScript version.

And, because the JS world is a bit nuts, you don't need to wait! We can teach Babel how to understand this syntax. Copy the package name, find your terminal and run:

yarn add babel-plugin-transform-object-rest-spread --dev
Oh, and as I mentioned before, most of these Babel plugins will have a slightly new name in the future: @babel/plugin-transform-object-rest-spread. But, it's really the same library.