38.5....Object Rest Spread{Using the Object Rest Spread}
Using the Object Rest Spread

Let's finish this! Down in catch, remove Object.assign(), remove the second argument and add ...newState.

And one more time: scroll down to handleDeleteRepLog(). We don't need this weird code anymore! Just return a new object with ...repLog then isDeleting: true.

I love that. And even better, when we refresh, it's not broken! We rock!