40.5.Passing Server Data to React Props{Dumping JavaScript in Twig}
Heck, we can add withHeart => true... because I like the heart.

Remove the die and pass this into twig as a new repLogAppProps variable.

Ready for the last piece? Delete the old JavaScript object and replace it with: {{ repLogAppProps|json_encode|raw }}.

That will print that array as JSON... which of course, is the same as JavaScript.

Ah, do you love it? It's now very easy to pass dynamic values or initial state into your app. Try it: refresh!