41.5.Reusable Components{Merging Prop Values in your Component}
Resolve undefined class of the button in previous commit. First, add Button.propTypes to advertise that we accept a className prop that's a string. We could make this required... or we can allow it to be optional, but fix that undefined problem. To do that, set Button.defaultProps to an object with className set to empty quotes.

Problem solved! Try it again. Wait! What? Now the class attribute is empty? How is that even possible?