41.6.Reusable Components{Merging Prop Values in your Component}
To see why class attribute was empty in previous commit, go back to RepLogCreator and pass a className prop here: btn-primary.

Go refresh again. Huh, now it has that class... but not the btn class. Here's the deal: sure, we have this className prop here. But, thanks to the ...props, the className prop we're passing in overrides the first one! We could move the ...props first, but, in general, we do want to allow whoever uses this component to override its props.